% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/snpData.R
\docType{methods}
\name{allele.freq,snpData-method}
\alias{allele.freq,snpData-method}
\title{Allele frequency}
\usage{
\S4method{allele.freq}{snpData}(x)
}
\arguments{
\item{x}{snpData object}
}
\description{
Allele frequency
}

