% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGeneric.R
\docType{methods}
\name{allele.freq}
\alias{allele.freq}
\title{Frequency of alleles in data.}
\usage{
allele.freq(x)
}
\arguments{
\item{x}{R object.}
}
\description{
Frequency of alleles in data.
}

