% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{alleleFreq}
\alias{alleleFreq}
\title{Allele Frequency.}
\usage{
alleleFreq(x, maf = TRUE)
}
\arguments{
\item{x}{an object, for which a corresponding method exists.}

\item{maf}{logical, if true minor allele frequency (default), other ways allele
frequency of 'allele.1'.}
}
\description{
Frequency of alleles in data set.
}

