% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpData.R
\docType{methods}
\name{as.matrix,snpData-method}
\alias{as.matrix,snpData-method}
\title{snpData as matrix}
\usage{
\S4method{as.matrix}{snpData}(x)
}
\arguments{
\item{x}{An object of class \linkS4class{snpData}.}
}
\description{
Matrix from an object of class \linkS4class{snpData}.
}
\examples{
# file containing example data for SNP data
gfile <- system.file("extdata/snpdata.csv", package = "qtcat")
snp <- read.snpData(gfile, sep = ",")
snpmat <- as.matrix(snp)

}

