% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpCluster.R
\name{clarans}
\alias{clarans}
\title{K-medoids clustering of SNPs using randomized search}
\usage{
clarans(snp, k, maxNeigbours = 100, nLocal = 10, mc.cores = 1)
}
\arguments{
\item{snp}{an object of class \linkS4class{snpMatrix}.}

\item{k}{a positive integer specifying the number of clusters, has to be greater than
one and less than the number of SNPs.}

\item{maxNeigbours}{a positive integer specifying the maximum number of randomized
searches.}

\item{nLocal}{a positive integer specifying the number of optimisation runs.}

\item{mc.cores}{a positive integer for the number of cores for parallel computing. See
\code{\link[parallel]{mclapply}} for details.}
}
\description{
Partitioning (clustering) into k clusters "around medoids" by randomized
search. \code{1-abs(cor)} is used as distance between SNPs.
}
\details{
The K-medoids clustering is implemented as clustering large applications based
upon randomized search (CLARANS) algorithm (Ng and Han 2002). CLARANS is a modification
of the partitioning around medoids (PAM) algorithm \code{\link[cluster]{pam}}. Where the
PAM algorithm is estimating all distances between SNPs and the respective medoids,
CLARANS is searching a random subset of the SNPs. This is independently repeated several
times and the result which minimises the average distance the most is reported. This
produces results close to those of the PAM algorithm (Ng and Han 2002), though the
number of runs and the subset size have to be arbitrarily chosen by the user. The
algorithm has two advantages: (i) the number of distance comparisons is dramatically
reduced; and (ii) parallelizing is straightforward.
}
\examples{
# file containing example data for SNP data
gfile <- system.file("extdata/snpdata.csv", package = "qtcat")
snp <- read.snpData(gfile, sep = ",")

clust <- clarans(snp, 3)

}
\references{
Ng and J. Han (2002). CLARANS: A method for clustering objects for spatial
data mining. \emph{IEEE Transactions on Knowledge and Data Engineering}.
\url{http://dx.doi.org/10.1109/TKDE.2002.1033770}).
}
