% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpCluster.R
\name{cutClust}
\alias{cutClust}
\title{Cut a qtcatClust object}
\usage{
cutClust(snp, snpClust, absCor = 1)
}
\arguments{
\item{snp}{an object of class \linkS4class{snpMatrix}.}

\item{snpClust}{an object of class \code{\link{qtcatClust}}.}

\item{absCor}{a cutting height in absolute value of correlation.}
}
\description{
Cut a qtcatClust object at an specific height.
}
\examples{
# file containing example data for SNP data
gfile <- system.file("extdata/snpdata.csv", package = "qtcat")
snp <- read.snpData(gfile, sep = ",")
clust <- qtcatClust(snp)

cclust <- cutClust(snp, clust, .5)

}
