% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/snpCluster.R
\name{cutDend}
\alias{cutDend}
\title{Cut a dendrogram}
\usage{
cutDend(x, absCor)
}
\arguments{
\item{x}{A dendrogram.}

\item{absCor}{Cutting height in absolute value of correlation.}
}
\description{
Cut a dendrogram at a spcific hight.
}
\examples{
# file containing example data for SNP data
gfile <- system.file("extdata/snpdata.csv", package = "qtcat")
snp <- read.snpData(gfile, sep = ",")
clust <- qtcatClust(snp)
cdend <- cutDend(clust$dendrogram, .5)
}

