% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpCluster.R
\name{distCor}
\alias{distCor}
\title{Correlation based distance among SNPs}
\usage{
distCor(snp)
}
\arguments{
\item{snp}{An object of class \linkS4class{snpMatrix}.}
}
\description{
This function computes and returns a distance matrix. Distance is estimated
as one minus absolute value of the correlation coefficient 1-abs(cor). The estimation is
computed for all pairwise combinations SNPs.
}
\details{
See \code{\link[stats]{dist}} for details about the output object.
}
\examples{
# file containing example data for SNP data
gfile <- system.file("extdata/snpdata.csv", package = "qtcat")
snp <- read.snpData(gfile, sep = ",")

dist <- distCor(snp[, 1:10])

}
\seealso{
\code{\link[stats]{dist}}
}

