% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/snpData.R
\docType{methods}
\name{getPos,snpData-method}
\alias{getPos,snpData-method}
\title{Get position from snpData}
\usage{
\S4method{getPos}{snpData}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{snpData}.}
}
\description{
Genetic position info from an object of class
 \linkS4class{snpData}.
}
\examples{
# file containing example data for SNP data
gfile <- system.file("extdata/snpdata.csv", package = "qtcat")
snp <- read.snpData(gfile, sep = ",
pos <- getPos(snp)
}

