% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGeneric.R
\docType{methods}
\name{het.freq}
\alias{het.freq}
\title{Frequency of heterozygosity in data.}
\usage{
het.freq(x, dim = 1L)
}
\arguments{
\item{x}{data object.}

\item{dim}{dimension over which heterozygosity is estimated.}
}
\description{
Frequency of heterozygosity in data.
}

