% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/snpData.R
\docType{methods}
\name{hetFreq,snpData-method}
\alias{hetFreq,snpData-method}
\title{Heterozygosity}
\usage{
\S4method{hetFreq}{snpData}(x, dim = c(1, 2))
}
\arguments{
\item{x}{snpData object.}

\item{dim}{Interger for dimension.}
}
\description{
Heterozygosity
}
\examples{
# file containing example data
file <- system.file("extdata", "snpdata.csv", package = "qtcat")
snp <- read.snpData(file, sep = ",")
hf1 <- hetFreq(snp, 1)
hf2 <- hetFreq(snp, 2)
}

