% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGeneric.R
\docType{methods}
\name{hetFreq}
\alias{hetFreq}
\title{Frequency of heterozygosity in data.}
\usage{
hetFreq(x, dim = 1L)
}
\arguments{
\item{x}{R object.}

\item{dim}{Dimension over which heterozygosity is estimated.}
}
\description{
Frequency of heterozygosity in data.
}

