% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpCluster.R
\name{identicals}
\alias{identicals}
\title{Perfect simiarity  SNP clusters}
\usage{
identicals(snp, mc.cores = 1)
}
\arguments{
\item{snp}{An object of class \linkS4class{snpData}.}

\item{mc.cores}{A positive integer for the number of cores for parallel computing. See 
\code{\link[parallel]{mclapply}} for details.}
}
\description{
Finds perfect similarity cluster of SNPs. This is specially us full in 
crossed populations.
}
\examples{
# file containing example data for SNP data
gfile <- system.file("extdata/snpdata.csv", package = "qtcat")
snp <- read.snpData(gfile, sep = ",")

ident <- identicals(snp)

}

