% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpImpute.R
\name{imputeMedo}
\alias{imputeMedo}
\title{Impute missing information at medoid SNPs}
\usage{
imputeMedo(SNP, clust, hier, flipAlleles, min.absCor = 0.25)
}
\arguments{
\item{clust}{A named vector of clusters.}

\item{hier}{A object of class hierarchy.}

\item{min.absCor}{A minimum value of correlation. If missing values still exist if this
point in the hierarchy is reached, imputing is done via allele frequencies.}

\item{snp}{An object of class \linkS4class{snpData}.}
}
\description{
Uses neighboring SNPs in the clustering hierarchy to impute alleles to
positions with missing values at medoid SNPs.
}
\keyword{internal}

