% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpImpute.R
\name{imputeMedoids}
\alias{imputeMedoids}
\title{Impute missing information at medoid SNPs}
\usage{
imputeMedoids(snp, snpClust, hier, min.absCor = 0.25, mc.cores = 1)
}
\arguments{
\item{snp}{an object of class \linkS4class{snpMatrix}.}

\item{snpClust}{an object of class \code{\link{qtcatClust}}.}

\item{hier}{an object of class hierarchy.}

\item{min.absCor}{a minimum value of correlation. If missing values still exist if this
point in the hierarchy is reached, imputing is done via allele frequencies.}

\item{mc.cores}{a number of cores for parallelising. Theoretical maximum is
\code{'B'}. For details see \code{\link[parallel]{mclapply}}.}
}
\description{
Uses neighboring SNPs in the clustering hierarchy to impute alleles to
positions with missing values at medoid SNPs.
}
\keyword{internal}

