% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpImpute.R
\name{imputeSnpData}
\alias{imputeSnpData}
\title{Impute missing information at each SNP}
\usage{
imputeSnpData(snp, snpClust, min.absCor = 0.25)
}
\arguments{
\item{snp}{An object of class \linkS4class{snpMatrix}.}

\item{snpClust}{An object of class \code{\link{qtcatClust}}.}

\item{min.absCor}{A minimum value of correlation. If missing values still exist if this
point in the hierarchy is reached, imputing is done via allele frequencies.}
}
\description{
Uses neighboring SNPs in the clustering hierarchy to impute alleles to
positions with missing values.
}

