% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpImpute.R
\name{imputeSnpIter}
\alias{imputeSnpIter}
\title{Impute missing information at a medoid SNPs from a group of neighbors}
\usage{
imputeSnpIter(snp, snpOfInt, inxSnpsToComp, snpOfIntFlip, flipAlleles)
}
\arguments{
\item{snp}{an object of class \linkS4class{snpMatrix}.}

\item{snpOfInt}{a vertor of the snp of interest.}

\item{inxSnpsToComp}{a index of neighbors.}

\item{snpOfIntFlip}{flip status of the snp of interest.}

\item{flipAlleles}{a vertor of telling for each SNP if allele one has allele freq. > 0.5
or not.}

\item{min.absCor}{a minimum value of correlation. If missing values still exist if this
point in the hierarchy is reached, imputing is done via allele frequencies.}
}
\description{
Uses neighboring SNPs in the clustering hierarchy to impute as many as
possible alleles to positions with missing values at medoid SNPs.
}
\keyword{internal}

