% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpImpute.R
\name{imputeSnpMatrix}
\alias{imputeSnpMatrix}
\title{Impute missing information at each SNP}
\usage{
imputeSnpMatrix(snp, snpClust, min.absCor = 0.25, mc.cores = 1)
}
\arguments{
\item{snp}{An object of class \linkS4class{snpMatrix}.}

\item{snpClust}{An object of class \code{\link{qtcatClust}}.}

\item{min.absCor}{A minimum value of correlation. If missing values still exist if this
point in the hierarchy is reached, imputing is done via allele frequencies.}

\item{mc.cores}{Number of cores for parallelising. Theoretical maximum is
\code{'B'}. For details see \code{\link[parallel]{mclapply}}.}
}
\description{
Uses neighboring SNPs in the clustering hierarchy to impute alleles to
positions with missing values.
}

