% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associationTest.R
\name{medoidsLm}
\alias{medoidsLm}
\title{Fitting a Linear Model to medoids}
\usage{
medoidsLm(object, pheno, geno, alpha = 0.05, min.absCor = 0.05)
}
\arguments{
\item{object}{an object of class \code{\link{qtcatHit}}.}

\item{pheno}{an object of class \code{\link{qtcatPheno}}.}

\item{geno}{an object of class \code{\link{qtcatGeno}}.}

\item{alpha}{an alpha level for significance estimation.}

\item{min.absCor}{minimum absolute value of correlation considered.}
}
\description{
Linear model between phenotype and medoids of QTCs (significant SNP
clusters).
}
\examples{
# If you want to run the examples, use:
# example(medoidsLm, run.dontrun = TRUE)
\dontrun{
# files containing example data for SNP data and the phenotype
pfile <- system.file("extdata/phenodata.csv", package = "qtcat")
gfile <- system.file("extdata/snpdata.csv", package = "qtcat")
pdat <- read.csv(pfile, header = TRUE)
snp <- read.snpData(gfile, sep = ",")
clust <- qtcatClust(snp)
geno <- qtcatGeno(snp, clust)
pheno <- qtcatPheno(names = pdat[, 1],
                    pheno = pdat[, 2],
                    covariates = model.matrix(~ pdat[, 3]))
fitted <- qtcatHit(pheno, geno)

# fitting a LM to the phenotype and QTC medoids
lmfitted <- medoidsLm(fitted, pheno, geno)
}

}

