% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associationTest.R
\name{medoQtc}
\alias{medoQtc}
\title{Find medoids of QTCs}
\usage{
medoQtc(object, geno, alpha = 0.05, min.absCor = 0.05)
}
\arguments{
\item{object}{an object of class \code{\link{qtcatHit}}.}

\item{geno}{an object of class \code{\link{qtcatGeno}}.}

\item{alpha}{an alpha level for significance estimation.}

\item{min.absCor}{minimum absolute value of correlation considered.}
}
\description{
Find a medoid of for each quantitative trait cluster (QTC).
}
\examples{
# If you want to run the examples, use:
# example(medoQtc, run.dontrun = TRUE)
\dontrun{
# files containing example data for SNP data and the phenotype
pfile <- system.file("extdata/phenodata.csv", package = "qtcat")
gfile <- system.file("extdata/snpdata.csv", package = "qtcat")
pdat <- read.csv(pfile, header = TRUE)
snp <- read.snpData(gfile, sep = ",")
clust <- qtcatClust(snp)
geno <- qtcatGeno(snp, clust)
pheno <- qtcatPheno(names = pdat[, 1],
                    pheno = pdat[, 2],
                    covariates = model.matrix(~ pdat[, 3]))
fitted <- qtcatHit(pheno, geno)

# QTC medoids
medo <- medoQtc(fitted, geno)
}

}

