% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpData.R
\docType{methods}
\name{naFreq,snpData-method}
\alias{naFreq,snpData-method}
\title{NA frequency}
\usage{
\S4method{naFreq}{snpData}(x, dim = 1)
}
\arguments{
\item{x}{An object of class \linkS4class{snpData}.}

\item{dim}{Integer for dimension.}
}
\description{
NA frequency in an object of class \linkS4class{snpData}.
}
\examples{
# file containing example data for SNP data
gfile <- system.file("extdata/snpdata.csv", package = "qtcat")
snp <- read.snpData(gfile, sep = ",")
na1 <- naFreq(snp, 1)
na2 <- naFreq(snp, 2)

}

