% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpData.R
\docType{methods}
\name{naFreq,snpMatrix-method}
\alias{naFreq,snpMatrix-method}
\title{NA frequency}
\usage{
\S4method{naFreq}{snpMatrix}(x, dim = 1)
}
\arguments{
\item{x}{an object of class \linkS4class{snpMatrix}.}

\item{dim}{a integer for dimension. 1 (default) is for rows (individuals), 2 is for
columns (SNPs).}
}
\description{
NA frequency in an object of class \linkS4class{snpMatrix}.
}
\examples{
# file containing example data for SNP data
gfile <- system.file("extdata/snpdata.csv", package = "qtcat")
snp <- read.snpData(gfile, sep = ",")
na1 <- naFreq(snp, 1)
na2 <- naFreq(snp, 2)

}

