% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associationTest.R
\name{plotQtc}
\alias{plotQtc}
\title{Plot resulting QTCs of the Hierarchical Inference Test}
\usage{
plotQtc(x, alpha = 0.05, xlab = "Chromosomes",
  ylab = expression(-log[10](italic(p))), col.axis = NULL, ...)
}
\arguments{
\item{x}{an object of class \code{\link{qtcatHit}}.}

\item{alpha}{an alpha level for significance estimation.}

\item{xlab}{a title for the x axis.}

\item{ylab}{a title for the y axis.}

\item{col.axis}{colors for axis line, tick marks, and title respectively.}

\item{...}{other graphical parameters may also be passed as arguments to this function.}
}
\description{
Plot the QTCs (significant cluster of SNPs) at their
position at the genome.
}
\examples{
# If you want to run the examples, use:
# example(plotQtc, run.dontrun = TRUE)
\dontrun{
# files containing example data for SNP data and the phenotype
pfile <- system.file("extdata/phenodata.csv", package = "qtcat")
gfile <- system.file("extdata/snpdata.csv", package = "qtcat")
pdat <- read.csv(pfile, header = TRUE)
snp <- read.snpData(gfile, sep = ",")
clust <- qtcatClust(snp)
geno <- qtcatGeno(snp, clust)
pheno <- qtcatPheno(names = pdat[, 1],
                    pheno = pdat[, 2],
                    covariates = model.matrix(~ pdat[, 3]))
fitted <- qtcatHit(pheno, geno)

# Plot the QTCs (loci37, loci260, and loci367 are causal)
plotQtc(fitted)
}

}

