% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/snpCluster.R
\name{qtcatClust}
\alias{qtcatClust}
\title{A three step approximated hirachical clustering among SNPs}
\usage{
qtcatClust(x, k, identicals = TRUE, maxNeigbours = 100, nLocal = 10,
  method = "complete", mc.cores = 1, trace = FALSE, ...)
}
\arguments{
\item{x}{A object of class \linkS4class{snpData}.}

\item{k}{A positive integer specifying the number of clusters, less than
the number of observations.}

\item{identicals}{Logical, if zero clustering.}

\item{maxNeigbours}{Positive integer, specifying the maximum number of
randomized searches.}

\item{nLocal}{Positive integer, specifying the number of optimisation runs.
Columns have to be similar to \code{x}.}

\item{method}{See hclust.}

\item{mc.cores}{Number of cores for parallel computing. See \code{mclapply}
in package parallel for details.}

\item{trace}{If \code{TRUE} it prints current status of the program.}

\item{...}{additional agruments for \code{\link[stats]{hclust}}}
}
\description{
Hirachical clustering for big data.
}
\seealso{
qtcatClarans
}

