% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/snpCluster.R
\name{qtcatCutDend}
\alias{qtcatCutDend}
\title{cluster dendrogram}
\usage{
qtcatCutDend(x, absCor)
}
\arguments{
\item{x}{dendrogram.}

\item{absCor}{cutting height in absolute value of correlation.}
}
\description{
cluster dendrogram
}
\examples{
# file containing example data
file <- system.file("extdata", "snpdata.csv", package = "qtcat")
snp <- read.snpData(file, sep = ",")
clust <- qtcatClust(snp)
cdend <- qtcatCutDend(clust$dendrogram, .5)
}

