% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/snpCluster.R
\name{qtcatDist}
\alias{qtcatDist}
\title{Correlation based distance among SNPs}
\usage{
qtcatDist(x)
}
\arguments{
\item{x}{An object of class \linkS4class{snpData}.}
}
\description{
This function computes and returns a distance matrix.
Distance is estimated as one minus absolute value of the correlation
coefficient 1-abs(cor). The estimation is computed for all pairwise
combinations SNPs.
}
\details{
See \code{\link[stats]{dist}} for details about the output object.
}
\seealso{
\code{\link[stats]{dist}}
}

