% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/associationTest.R
\name{qtcatGeno}
\alias{qtcatGeno}
\title{geno data}
\usage{
qtcatGeno(x, qtcatClust, height, max.height)
}
\arguments{
\item{x}{snpData}

\item{qtcatClust}{clusteing}

\item{height}{height}

\item{max.height}{max.height}
}
\description{
geno data
}

