% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/corCluster.R
\name{qtcatIdenticals}
\alias{qtcatIdenticals}
\title{Internal function}
\usage{
qtcatIdenticals(x, mc.cores = 1)
}
\arguments{
\item{x}{A object of class \linkS4class{snpData}.}

\item{mc.cores}{A positive integer for the number of cores for parallel
computing. See \code{\link[parallel]{mclapply}} for details.}
}
\description{
cluster at distance zero in qtcatClust
}

