% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/associationTest.R
\name{qtcatLM}
\alias{qtcatLM}
\title{Linear model}
\usage{
qtcatLM(x, pheno, geno, alpha = 0.05, min.absCor)
}
\arguments{
\item{x}{hit object.}

\item{pheno}{qtcatPheno object.}

\item{geno}{qtcatGeno object.}

\item{alpha}{Alpha level.}

\item{min.absCor}{Minimum absolute value of correlation considered.}
}
\description{
Linear model between phenotype amd medoids of significent SNP
clusters.
}

