% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/snpCluster.R
\name{qtcatMedoids}
\alias{qtcatMedoids}
\title{Estimates the medoids}
\usage{
qtcatMedoids(x, clusters)
}
\arguments{
\item{x}{An object of class \linkS4class{snpData}.}

\item{clusters}{Vector of cluster groups \code{snpData}.}
}
\description{
Estimates clusters and medoids.
}
\examples{
# file containing example data
file <- system.file("extdata", "snpdata.csv", package = "qtcat")
snp <- read.snpData(file, sep = ",")
clust <- qtcatClust(snp)
cclust <- qtcatCutClust(clust, .5)
mclust <- qtcatMedoids(snp, cclust)
}

