% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associationTest.R
\name{qtcatQtc}
\alias{qtcatQtc}
\title{Summarize results of Hierarchical Inference Test}
\usage{
qtcatQtc(object, alpha = 0.05, min.absCor = 0.05)
}
\arguments{
\item{object}{An object of class \code{\link{qtcatHit}}.}

\item{alpha}{An alpha level for significance estimation.}

\item{min.absCor}{Minimum absolute value of correlation considered.}
}
\description{
Summarizing the QTCs (significant cluster of SNPs) and their
position at the genome.
}
\examples{
# If you want to run the examples, use:
# example(qtcatQtc, run.dontrun = TRUE)
\dontrun{
# files containing example data for SNP data and the phenotype
pfile <- system.file("extdata/phenodata.csv", package = "qtcat")
gfile <- system.file("extdata/snpdata.csv", package = "qtcat")
pdat <- read.csv(pfile, header = TRUE)
snp <- read.snpData(gfile, sep = ",")
clust <- qtcatClust(snp)
geno <- qtcatGeno(snp, clust)
pheno <- qtcatPheno(names = pdat[, 1],
                    pheno = pdat[, 2],
                    covariates = model.matrix(~ pdat[, 3]))
fitted <- qtcatHit(pheno, geno)

# Summarizing the QTCs (loci37, loci260, and loci367 are causal)
qtcatQtc(fitted)
}

}

