% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpData.R
\name{read.snpData}
\alias{read.snpData}
\title{Read SNP tables as a snpMatrix object}
\usage{
read.snpData(file, sep = " ", quote = "\\"", na.string = "NA",
  nrows = -1L)
}
\arguments{
\item{file}{the name of the file which the data are to be read from. If it does not
contain an absolute path, the file name is relative to the current working directory,
\code{getwd()}. Tilde-expansion is performed where supported.}

\item{sep}{a field separator character. Values on each line of the file are separated
by this character.}

\item{quote}{the set of quoting characters. To disable quoting altogether, use
\code{quote = ""}.}

\item{na.string}{a string which is  interpreted as NA value.}

\item{nrows}{a integer, the maximum number of rows to read.}
}
\description{
Reads a file in table format and as a \linkS4class{snpMatrix} object.
}
\examples{
# file containing example data for SNP data
gfile <- system.file("extdata/snpdata.csv", package = "qtcat")
snp <- read.snpData(gfile, sep = ",")

}

