% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/corCluster.R
\name{sigClusters}
\alias{sigClusters}
\title{Include zero clusters x}
\usage{
sigClusters(x, clusters, alpha = 0.05, max.height)
}
\arguments{
\item{x}{hit object.}

\item{clusters}{object of class qtcatClust.}

\item{alpha}{alpha level.}

\item{max.height}{max. height to consider.}
}
\description{
Include zero clusters
}

