% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associationTest.R
\docType{methods}
\name{snpInfo,qtcatHit-method}
\alias{snpInfo,qtcatHit-method}
\title{Get position from qtcatHit object}
\usage{
\S4method{snpInfo}{qtcatHit}(object)
}
\arguments{
\item{object}{An object of class qtcatHit.}
}
\description{
Genetic position info from an object of class qtcatHit.
}

