% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpData.R
\docType{methods}
\name{snpInfo,snpData-method}
\alias{snpInfo,snpData-method}
\title{Get position from snpData}
\usage{
\S4method{snpInfo}{snpData}(object)
}
\arguments{
\item{object}{An object of class \linkS4class{snpData}.}
}
\description{
Genetic position info from an object of class
 \linkS4class{snpData}.
}
\examples{
# file containing example data for SNP data
gfile <- system.file("extdata/snpdata.csv", package = "qtcat")
snp <- read.snpData(gfile, sep = ",")
info <- snpInfo(snp)

}

