% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{snpInfo}
\alias{snpInfo}
\title{Extract genomic position and allele information.}
\usage{
snpInfo(object)
}
\arguments{
\item{object}{an object, for which a corresponding method exists.}
}
\description{
Extract genomic position and allele information from object.
}
