% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{snpMatrix-class}
\alias{snpMatrix-class}
\title{A S4 class to represent a SNP-matrix}
\description{
A S4 class to represent a SNP matrix. Storing SNP information, by using a
byte-level (raw) storage scheme, jointly with genomic position and allele information.
}
\section{Slots}{

\describe{
\item{\code{snpData}}{a matrix of SNPs stored in type 'raw'. 00 is NA, 01 homozygote AA, 02
heterozygote AB, and 03 homozygote BB.}

\item{\code{snpInfo}}{data.frame with four columns. The first col. contains the chromosomes,
the second col. the positions, the third col. the first allele and the fourth the second
allele.}

\item{\code{dim}}{an integer vector with exactly two non-negative values.}

\item{\code{dimnames}}{a list of length two; each component containing NULL or a character vector
length equal the corresponding dim element.}
}}

