% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/snpData.R
\docType{methods}
\name{[,snpData,ANY,ANY,missing-method}
\alias{[,snpData,ANY,ANY,missing-method}
\title{Sub snpData}
\usage{
\S4method{[}{snpData,ANY,ANY,missing}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{snpData object}

\item{i}{indices specifying elements to extract or replace. Indices are
numeric or character vectors.}

\item{j}{indices specifying elements to extract or replace. Indices are
numeric or character vectors}

\item{...}{not implemented}

\item{drop}{not implemented}
}
\description{
Sub snpData
}

