% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/snpData.R
\docType{methods}
\name{alleleFreq,snpData-method}
\alias{alleleFreq,snpData-method}
\title{Allele frequency}
\usage{
\S4method{alleleFreq}{snpData}(x)
}
\arguments{
\item{x}{snpData object.}
}
\description{
Allele frequency
}
\examples{
# file containing example data
file <- system.file("extdata", "snpdata.csv", package = "qtcat")
snp <- read.snpData(file, sep = ",")
af <- alleleFreq(snp)
}

