% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/snpData.R
\docType{methods}
\name{alleleFreq,snpData-method}
\alias{alleleFreq,snpData-method}
\title{Allele frequency}
\usage{
\S4method{alleleFreq}{snpData}(x)
}
\arguments{
\item{x}{An object of class \linkS4class{snpData}.}
}
\description{
Allele frequency an object of class \linkS4class{snpData}.
}
\examples{
# file containing example data for SNP data
gfile <- system.file("extdata/snpdata.csv", package = "qtcat")
snp <- read.snpData(gfile, sep = ",")
af <- alleleFreq(snp)
}

