% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpData.R
\docType{methods}
\name{alleleFreq,snpMatrix-method}
\alias{alleleFreq,snpMatrix-method}
\title{Allele frequency}
\usage{
\S4method{alleleFreq}{snpMatrix}(x, maf = TRUE)
}
\arguments{
\item{x}{an object of class \linkS4class{snpMatrix}.}

\item{maf}{logical, if true minor allele frequency (default), other ways allele
frequency of 'allele.1'.}
}
\description{
Allele frequency an object of class \linkS4class{snpMatrix}.
}
\examples{
# file containing example data for SNP data
gfile <- system.file("extdata/snpdata.csv", package = "qtcat")
snp <- read.snpData(gfile, sep = ",")
af <- alleleFreq(snp)

}

