% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{alleleFreq}
\alias{alleleFreq}
\title{Frequency of alleles in data.}
\usage{
alleleFreq(x, maf = TRUE)
}
\arguments{
\item{x}{An object.}

\item{maf}{If true minor allele frequency (default), other ways allele frequency of 
'allele.1'.}
}
\description{
Frequency of alleles in data.
}

