% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/snpData.R
\docType{methods}
\name{as.matrix,snpData-method}
\alias{as.matrix,snpData-method}
\title{snpData as matrix}
\usage{
\S4method{as.matrix}{snpData}(x, inx1 = 1:ncol(x), inx2 = NULL, ...)
}
\arguments{
\item{x}{snpData object}

\item{inx1}{...}

\item{inx2}{...}

\item{...}{not implemented}
}
\description{
snpData as matrix
}

