% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpData.R
\docType{methods}
\name{as.matrix,snpMatrix-method}
\alias{as.matrix,snpMatrix-method}
\title{snpMatrix as matrix}
\usage{
\S4method{as.matrix}{snpMatrix}(x)
}
\arguments{
\item{x}{An object of class \linkS4class{snpMatrix}.}
}
\description{
Matrix from an object of class \linkS4class{snpMatrix}.
}
\examples{
# file containing example data for SNP data
gfile <- system.file("extdata/snpdata.csv", package = "qtcat")
snp <- read.snpData(gfile, sep = ",")
snpmat <- as.matrix(snp)

}

