% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/snpData.R
\name{as.snpData}
\alias{as.snpData}
\title{converts data to  snpData object}
\usage{
as.snpData(x, position, alleleCoding = c(-1, 0, 1), alleles = NULL)
}
\arguments{
\item{x}{matrix with indoviduals in rows and SNPs in columns.}

\item{position}{matrix with chromosome and position in rows and
SNPs in columns.}

\item{alleleCoding}{coding of \code{x} for hom het hom.}

\item{alleles}{labels of alleles, for each SNP.}
}
\description{
converts data to  snpData object
}

