% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpCluster.R
\name{cutDend}
\alias{cutDend}
\title{Cut a dendrogram}
\usage{
cutDend(snp, dend, absCor = 1)
}
\arguments{
\item{snp}{A object of class \linkS4class{snpData}.}

\item{dend}{A dendrogram.}

\item{absCor}{Cutting height in absolute value of correlation.}
}
\description{
Cut a dendrogram at a spcific hight.
}
\examples{
# file containing example data for SNP data
gfile <- system.file("extdata/snpdata.csv", package = "qtcat")
snp <- read.snpData(gfile, sep = ",")
clust <- qtcatClust(snp)

cdend <- cutDend(snp, clust$dendrogram, .5)

}

