% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpCluster.R
\name{distCor}
\alias{distCor}
\title{Correlation based distance between SNPs}
\usage{
distCor(snp)
}
\arguments{
\item{snp}{an object of class \linkS4class{snpMatrix}.}
}
\description{
This function computes a distance matrix. The distance is estimated
as one minus the absolute value of the correlation coefficient \code{1 - abs(cor)}.
}
\details{
See \code{\link[stats]{dist}} for details about the output object.
}
\examples{
# file containing example data for SNP data
gfile <- system.file("extdata/snpdata.csv", package = "qtcat")
snp <- read.snpData(gfile, sep = ",")

dist <- distCor(snp[, 1:10])

}
\seealso{
\code{\link[stats]{dist}}
}

