% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/snpData.R
\docType{methods}
\name{getPos,snpData-method}
\alias{getPos,snpData-method}
\title{Get position from snpData}
\usage{
\S4method{getPos}{snpData}(object)
}
\arguments{
\item{object}{snpData object.}
}
\description{
Get position from snpData
}
\examples{
# file containing example data
file <- system.file("extdata", "snpdata.csv", package = "qtcat")
snp <- read.snpData(file, sep = ",")
pos <- getPos(snp)
}

