% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllGeneric.R
\docType{methods}
\name{getPos}
\alias{getPos}
\title{Extract position information from object.}
\usage{
getPos(object)
}
\arguments{
\item{object}{R object.}
}
\description{
Extract position information from object.
}

