% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/snpData.R
\docType{methods}
\name{het.freq,snpData-method}
\alias{het.freq,snpData-method}
\title{Heterozygosity}
\usage{
\S4method{het.freq}{snpData}(x, dim = c(1, 2))
}
\arguments{
\item{x}{snpData object}

\item{dim}{interger for dimention}
}
\description{
Heterozygosity
}

