% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/snpData.R
\docType{methods}
\name{hetFreq,snpData-method}
\alias{hetFreq,snpData-method}
\title{Heterozygosity}
\usage{
\S4method{hetFreq}{snpData}(x, dim = c(1, 2))
}
\arguments{
\item{x}{An object of class \linkS4class{snpData}.}

\item{dim}{Integer for dimension.}
}
\description{
Heterozygosity an object of class \linkS4class{snpData}.
}
\examples{
# file containing example data for SNP data
gfile <- system.file("extdata/snpdata.csv", package = "qtcat")
snp <- read.snpData(gfile, sep = ",
hf1 <- hetFreq(snp, 1)
hf2 <- hetFreq(snp, 2)
}

