% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{hetFreq}
\alias{hetFreq}
\title{Heterozygosity Frequency.}
\usage{
hetFreq(x, dim = 1L)
}
\arguments{
\item{x}{an object, for which a corresponding method exists.}

\item{dim}{dimension over which heterozygosity is estimated. 1 (default) is for rows
(individuals), 2 is for columns (SNPs).}
}
\description{
Frequency of heterozygosity in data set.
}
