% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{hetFreq}
\alias{hetFreq}
\title{Frequency of heterozygosity in data.}
\usage{
hetFreq(x, dim = 1L)
}
\arguments{
\item{x}{An object.}

\item{dim}{Dimension over which heterozygosity is estimated.}
}
\description{
Frequency of heterozygosity in data.
}

