% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpImpute.R
\name{imputeMedo}
\alias{imputeMedo}
\title{Impute missing information at medoid SNPs}
\usage{
imputeMedo(snp, clust, hier, min.absCor = 0.25)
}
\arguments{
\item{snp}{An object of class \linkS4class{snpMatrix}.}

\item{clust}{A named vector of clusters.}

\item{hier}{A object of class hierarchy.}

\item{min.absCor}{A minimum value of correlation. If missing values still exist if this
point in the hierarchy is reached, imputing is done via allele frequencies.}
}
\description{
Uses neighboring SNPs in the clustering hierarchy to impute alleles to
positions with missing values at medoid SNPs.
}
\keyword{internal}

