% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpImpute.R
\name{imputeSnp}
\alias{imputeSnp}
\title{Impute missing information at a medoid SNPs from a group of neighbors}
\usage{
imputeSnp(snp, snpOfInt, inxSnpsToComp, snpOfIntFlip, flipAlleles)
}
\arguments{
\item{snp}{An object of class \linkS4class{snpMatrix}.}

\item{snpOfInt}{A vertor of the snp of interest.}

\item{inxSnpsToComp}{Index of neighbors.}

\item{snpOfIntFlip}{Flip status of the snp of interest.}

\item{flipAlleles}{A vertor of telling for each SNP if allele one has allele freq. > 0.5
or not.}

\item{min.absCor}{A minimum value of correlation. If missing values still exist if this
point in the hierarchy is reached, imputing is done via allele frequencies.}
}
\description{
Uses neighboring SNPs in the clustering hierarchy to impute as many as
possible alleles to positions with missing values at medoid SNPs.
}
\keyword{internal}

