% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/snpCluster.R
\name{medoids}
\alias{medoids}
\title{Find medoids of SNP clusters}
\usage{
medoids(snp, snpClust, absCor)
}
\arguments{
\item{snp}{An object of class \linkS4class{snpData}.}

\item{snpClust}{An object of class \code{\link{qtcatClust}}.}

\item{absCor}{cutting height in absolute value of correlation.}
}
\description{
Find the medoid of each SNP cluster.
}
\examples{
# file containing example data for SNP data
gfile <- system.file("extdata/snpdata.csv", package = "qtcat")
snp <- read.snpData(gfile, sep = ",")
clust <- qtcatClust(snp)

mclust <- medoids(snp, clust, .5)
}

