% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{naFreq}
\alias{naFreq}
\title{Frequency of NAs in data.}
\usage{
naFreq(x, dim = 1L)
}
\arguments{
\item{x}{An object.}

\item{dim}{Dimension over which heterozygosity is estimated.}
}
\description{
Frequency of NAs in data.
}

