% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/qtcatPackage.R
\docType{package}
\name{qtcat-package}
\alias{qtcat}
\alias{qtcat-package}
\title{Quantitative trait cluster association test}
\description{
Quantitative trait cluster association test
}
\author{
Jonas Klasen
}

