% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qtcatPackage.R
\docType{package}
\name{qtcat-package}
\alias{qtcat}
\alias{qtcat-package}
\title{Quantitative Trait Cluster Association Test}
\description{
An association mapping method which jointly analyses all SNPs
at once and at the same time accounts for the correlation between them. This
makes correction for population structure unnecessary and therefore
increases power compared to classical methods like the mixed model.
}
\author{
Jonas Klasen
}

