% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtcatPackage.R
\docType{package}
\name{qtcat-package}
\alias{qtcat-package}
\alias{qtcat}
\title{Quantitative Trait Cluster Association Test}
\description{
All SNPs are jointly associated to the phenotype and at the same time
correlation among them is considered. Thus, correction for population structure becomes
unnecessary, which in many cases results in a power advantages compared to other methods.
}
\author{
Jonas Klasen
}
