% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpCluster.R
\name{qtcatClust}
\alias{qtcatClust}
\title{Hierarchical clustering for big SNP data sets.}
\usage{
qtcatClust(snp, k, identicals = TRUE, maxNeigbours = 100, nLocal = 10,
  method = "complete", mc.cores = 1, trace = FALSE, ...)
}
\arguments{
\item{snp}{an object of class \linkS4class{snpMatrix}.}

\item{k}{a positive integer specifying the number of clusters, less than the number of
observations.}

\item{identicals}{logical, if zero clustering.}

\item{maxNeigbours}{a positive integer, specifying the maximum number of randomized
searches.}

\item{nLocal}{a positive integer, specifying the number of optimisation runs. Columns
have to be similar to \code{snp}.}

\item{method}{see hclust.}

\item{mc.cores}{a number of cores for parallel computing. See \code{mclapply} in package
parallel for details.}

\item{trace}{logical, if \code{TRUE} it prints current status of the program.}

\item{...}{additional argruments for \code{\link[stats]{hclust}}}
}
\description{
A three step approximated hierarchical clustering of SNPs suitable to
large data sets.
}
\examples{
# file containing example data for SNP data
gfile <- system.file("extdata/snpdata.csv", package = "qtcat")
snp <- read.snpData(gfile, sep = ",")

clust <- qtcatClust(snp)

}
\seealso{
clarans
}

