% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/snpCluster.R
\name{qtcatCutClust}
\alias{qtcatCutClust}
\title{Cut a qtcatClust object}
\usage{
qtcatCutClust(x, absCor)
}
\arguments{
\item{x}{qtcatClust object.}

\item{absCor}{cutting height in absolute value of correlation.}
}
\description{
Cut a qtcatClust object at an specific height.
}
\examples{
# file containing example data for SNP data
gfile <- system.file("extdata/snpdata.csv", package = "qtcat")
snp <- read.snpData(gfile, sep = ",")
clust <- qtcatClust(snp)
cclust <- qtcatCutClust(clust, .5)
}

