% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/snpCluster.R
\name{qtcatCutClust}
\alias{qtcatCutClust}
\title{cluster qtcatClust object}
\usage{
qtcatCutClust(x, h)
}
\arguments{
\item{x}{qtcatClust object}

\item{h}{cutting hight}
}
\description{
cluster qtcatClust object
}

