% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/snpCluster.R
\name{qtcatCutClust}
\alias{qtcatCutClust}
\title{cluster qtcatClust object}
\usage{
qtcatCutClust(x, absCor)
}
\arguments{
\item{x}{qtcatClust object.}

\item{absCor}{cutting height in absolute value of correlation.}
}
\description{
cluster qtcatClust object
}

