% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/associationTest.R
\name{qtcatGeno}
\alias{qtcatGeno}
\title{geno data}
\usage{
qtcatGeno(x, clusters, height, max.height = 0.3)
}
\arguments{
\item{x}{snpData object.}

\item{clusters}{Clustering.}

\item{height}{Height.}

\item{max.height}{Max. height.}
}
\description{
geno data
}

