% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/associationTest.R
\name{qtcatGeno}
\alias{qtcatGeno}
\title{geno data}
\usage{
qtcatGeno(x, clusters, absCor, min.absCor = 0.7)
}
\arguments{
\item{x}{snpData object.}

\item{clusters}{Clustering.}

\item{absCor}{Vector of absolute value of correlations considered in the
hierarchy.}

\item{min.absCor}{Minimum absolute value of correlation considered.}
}
\description{
geno data
}
\examples{
# file containing example data
file <- system.file("extdata", "snpdata.csv", package = "qtcat")
snp <- read.snpData(file, sep = ",")
clust <- qtcatClust(snp)
geno <- qtcatGeno(snp, clust)
}

