% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/associationTest.R
\name{qtcatGeno}
\alias{qtcatGeno}
\title{geno data}
\usage{
qtcatGeno(x, clusters, height, max.height = 0.3)
}
\arguments{
\item{x}{snpData}

\item{clusters}{clusteing}

\item{height}{height}

\item{max.height}{max.height}
}
\description{
geno data
}

