% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associationTest.R
\name{qtcatGeno}
\alias{qtcatGeno}
\title{A genotype object constructor}
\usage{
qtcatGeno(snp, snpClust, absCor, min.absCor = 0.5, mc.cores = 1)
}
\arguments{
\item{snp}{an object of S4 class \linkS4class{snpMatrix}.}

\item{snpClust}{an object of class \code{\link{qtcatClust}}.}

\item{absCor}{a vector of absolute value of correlations considered in the hierarchy.}

\item{min.absCor}{a minimum absolute value of correlation which is considered. A value
in the range from 0 to 1.}

\item{mc.cores}{a number of cores for parallelising. The maximum is \code{'B'}. For
details see \code{\link[parallel]{mclapply}}.}
}
\description{
Constructs a S3-object containing a SNP design matrix and a hierarchy. If
a SNP in the input object contains missing data, the clustering is used to impute
information from highly correlated neighbor SNPs. The genotype object is needed for
\code{\link{qtcatHit}} as input.
}
\examples{
# file containing example data for SNP data
gfile <- system.file("extdata/snpdata.csv", package = "qtcat")
snp <- read.snpData(gfile, sep = ",")
clust <- qtcatClust(snp)

# Construct geotype object
geno <- qtcatGeno(snp, clust)

}

