% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/associationTest.R
\name{qtcatGeno}
\alias{qtcatGeno}
\title{A genotype object constructor}
\usage{
qtcatGeno(snp, snpClust, absCor, min.absCor = 0.7)
}
\arguments{
\item{snp}{An object of S4 class \linkS4class{snpData}.}

\item{snpClust}{An object of class \code{\link{qtcatClust}}.}

\item{absCor}{Vector of absolute value of correlations considered in the
hierarchy.}

\item{min.absCor}{Minimum absolute value of correlation considered. A value
in the range from 0 to 1.}
}
\description{
Constructs an S3 object containing SNP matrix and SNP
hierarchy. This is needed for \code{\link{qtcatHit}} as input.
}
\examples{
# file containing example data for SNP data
gfile <- system.file("extdata/snpdata.csv", package = "qtcat")
snp <- read.snpData(gfile, sep = ",")
clust <- qtcatClust(snp)

# Construct geotype object
geno <- qtcatGeno(snp, clust)
}

