% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/associationTest.R
\name{qtcatHit}
\alias{qtcatHit}
\title{hit}
\usage{
qtcatHit(pheno, geno, B = 50, p.samp1 = 0.5, gamma = seq(0.05, 0.99, by =
  0.01), max.p.esti = 1, mc.cores = 1L, trace = FALSE, ...)
}
\arguments{
\item{pheno}{qtcatPheno}

\item{geno}{qtcatGeno}

\item{B}{number of sample-splits.}

\item{p.samp1}{fraction of data used for the LASSO. The ANOVA uses
\code{1 - p.samp1}.}

\item{gamma}{vector of gamma-values.}

\item{max.p.esti}{maximum alpha level. All p-values above this value are set
to one. Small \code{max.p.esti} values reduce computing time.}

\item{mc.cores}{number of cores for parallelising. Theoretical maximum is
'B'. For details see \code{\link[parallel]{mclapply}}.}

\item{trace}{if TRUE it prints current status of the program.}

\item{...}{additional arguments for \code{\link[glmnet]{cv.glmnet}}.}
}
\description{
hit
}

