% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/associationTest.R
\name{qtcatLm}
\alias{qtcatLm}
\title{Fitting a Linear Model to medoids}
\usage{
qtcatLm(x, pheno, geno, alpha = 0.05, min.absCor)
}
\arguments{
\item{x}{An object of class \code{\link{qtcatHit}}.}

\item{pheno}{An object of class \code{\link{qtcatPheno}}.}

\item{geno}{An object of class \code{\link{qtcatGeno}}.}

\item{alpha}{An alpha level for significance estimation.}

\item{min.absCor}{Minimum absolute value of correlation considered.}
}
\description{
Linear model between phenotype and medoids of QTCs (significant
SNP clusters).
}
\examples{
# files containing example data for SNP data and the phenotype
pfile <- system.file("extdata/phenodata.csv", package = "qtcat")
gfile <- system.file("extdata/snpdata.csv", package = "qtcat")
pdat <- read.csv(pfile, header = TRUE)
snp <- read.snpData(gfile, sep = ",")
clust <- qtcatClust(snp)
geno <- qtcatGeno(snp, clust)
pheno <- qtcatPheno(pdat)
fitted <- qtcatHit(pheno, geno)
# fitting a LM to the phenotype and QTC medoids
lmfitted <- qtcatLm(fitted, pheno, geno)
}

