% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/snpCluster.R
\name{qtcatMedoids}
\alias{qtcatMedoids}
\title{Find the medoid}
\usage{
qtcatMedoids(x, clusters)
}
\arguments{
\item{x}{An object of class \linkS4class{snpData}.}

\item{clusters}{Vector of cluster groups \code{qtcatCutClust}.}
}
\description{
Find the medoid of each cluster.
}
\examples{
# file containing example data for SNP data
gfile <- system.file("extdata/snpdata.csv", package = "qtcat")
snp <- read.snpData(gfile, sep = ",")
clust <- qtcatClust(snp)
cclust <- qtcatCutClust(clust, .5)
mclust <- qtcatMedoids(snp, cclust)
}

