% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/snpCluster.R
\name{qtcatMedoids}
\alias{qtcatMedoids}
\title{Estimates the medoids.}
\usage{
qtcatMedoids(x, clusters)
}
\arguments{
\item{x}{A object of class \linkS4class{snpData}.}

\item{clusters}{vector of cluster groups \code{snpData}.}
}
\description{
Estimates clusters and medoids.
}

