% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/associationTest.R
\name{qtcatPheno}
\alias{qtcatPheno}
\title{peno of xxx}
\usage{
qtcatPheno(x)
}
\arguments{
\item{x}{data.frame fisrt column individual names secend column pheno
additional columns aditional variables.}
}
\description{
peno of xxx
}

