% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/associationTest.R
\name{qtcatPheno}
\alias{qtcatPheno}
\title{geno of xxx}
\usage{
qtcatPheno(x)
}
\arguments{
\item{x}{data.frame, first column individual names, second column phenotype.
Additional columns for additional variables.}
}
\description{
geno of xxx
}
\examples{
# file containing example data
file <- system.file("extdata", "phenodata.csv", package = "qtcat")
pdat <- read.csv(file, header = TRUE)
pheno <- qtcatPheno(pdat)
}

