% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/associationTest.R
\name{qtcatPheno}
\alias{qtcatPheno}
\title{A phenotype object constructor}
\usage{
qtcatPheno(names, pheno, family = "gaussian", covariates = NULL)
}
\arguments{
\item{names}{A vector of individual names of length 'n'.}

\item{pheno}{A vector of length 'n' or a matrix size 'n x 2' in case of binomial family. 
This contains the phenotypic observations.}

\item{family}{A character string specifying the family of the phenotype distribution. 
Either "gaussian" (default) or "binomial".}

\item{covariates}{A matrix typically generated by a call of
\code{\link[stats]{model.matrix}}. It contain additional variables influencing the 
phenotype e.g. environmental and experimental components.}
}
\description{
Constructs an S3 object containing phenotype and if additional covariats 
exist a design matrix of those. This is needed for \code{\link{qtcatHit}} as input.
}
\examples{
# file containing example data for a phenotype.
pfile <- system.file("extdata/phenodata.csv", package = "qtcat")
pdat <- read.csv(pfile, header = TRUE)

# Construct phenotype object
pheno <- qtcatPheno(names = pdat[, 1],
                    pheno = pdat[, 2],
                    covariates = model.matrix(~ pdat[, 3]))

}

