% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/associationTest.R
\name{qtcatQtc}
\alias{qtcatQtc}
\title{Summarize results of Hierarchical Inference Test}
\usage{
qtcatQtc(x, alpha = 0.05, min.absCor)
}
\arguments{
\item{x}{An object of class \code{\link{qtcatHit}}.}

\item{alpha}{An alpha level for significance estimation.}

\item{min.absCor}{Minimum absolute value of correlation considered.}
}
\description{
Summarizing the QTCs (significant cluster of SNPs) and their
position at the genome.
}
\examples{
# files containing example data for SNP data and the phenotype
pfile <- system.file("extdata/phenodata.csv", package = "qtcat")
gfile <- system.file("extdata/snpdata.csv", package = "qtcat")
pdat <- read.csv(pfile, header = TRUE)
snp <- read.snpData(gfile, sep = ",")
clust <- qtcatClust(snp)
geno <- qtcatGeno(snp, clust)
pheno <- qtcatPheno(pdat)
fitted <- qtcatHit(pheno, geno)
# Summarizing the QTCs
qtcatQtc(fitted)
}

