% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/associationTest.R
\name{qtcatSigClust}
\alias{qtcatSigClust}
\title{Include zero clusters x}
\usage{
qtcatSigClust(x, alpha = 0.05, min.absCor)
}
\arguments{
\item{x}{hit object.}

\item{alpha}{Alpha level.}

\item{min.absCor}{Minimum absolute value of correlation considered.}
}
\description{
Include zero clusters.
}

