% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpCluster.R
\name{rename.leafs}
\alias{rename.leafs}
\title{Rename dendrogram leafs}
\usage{
rename.leafs(dend, labels)
}
\arguments{
\item{dend}{a dendrogram.}

\item{labels}{a vector of new names.}
}
\description{
Rename dendrogram leafs.
}
\keyword{internal}

