% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{snpData-class}
\alias{snpData-class}
\title{Class "snpData" for a SNP matrix (S4 class)}
\description{
An S4 class for a SNP matrix. Storing SNP information using a
byte-level (raw) storage scheme jointly with positions.
}
\section{Slots}{

\describe{
\item{\code{snpData}}{A matrix of SNPs stored in type 'raw'.}

\item{\code{position}}{A matrix with two rows. The first row contains the
chromosomes the second row contains the positions.}

\item{\code{alleles}}{A matrix of allele labels: A, T, C or, G.}

\item{\code{dim}}{integer vector with exactly two non-negative values.}

\item{\code{dimnames}}{list of length two; each component containing NULL or a
character vector length equal the corresponding dim element.}
}}

