% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{snpInfo}
\alias{snpInfo}
\title{Extract position and allele information from object.}
\usage{
snpInfo(object)
}
\arguments{
\item{object}{An object.}
}
\description{
Extract position and allele information from object.
}

