% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpData.R
\docType{methods}
\name{[,snpMatrix,ANY,ANY,missing-method}
\alias{[,snpMatrix,ANY,ANY,missing-method}
\title{Sub snpMatrix}
\usage{
\S4method{[}{snpMatrix,ANY,ANY,missing}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{An object of class \linkS4class{snpMatrix}.}

\item{i}{Indices specifying elements to extract or replace. Indices are booleans,
numeric or character vectors.}

\item{j}{indices specifying elements to extract or replace. Indices are booleans,
numeric or character vectors.}

\item{...}{Not implemented.}

\item{drop}{Not implemented.}
}
\description{
Subsetting an object of class \linkS4class{snpMatrix}.
}

